keep just datetime_to_isostring and make_datetime on public interface
